Modify frontend to have better UX
Added a hack that adds christian to pop songs.. because one is
frequently misclassified as the other. Also made a simple jquery
check that does not keep appending to the probabilities table,
but only adds to the single row (or adds it the first time). same
for the predicted genre that comes written.